requirements.txt
streamlit==1.24.0
pinecone-client==2.2.2
cohere==3.3.0
transformers==4.30.0
torch==2.0.0
PyPDF2==3.0.0